🧹 [Code Health] Extract plugin architecture to smaller modules
🎯 What: Extracted the large plugins.py into a modular plugin package with exceptions.py, models.py, hooks.py, loader.py, registry.py, and __init__.py.
💡 Why: To improve readability and maintainability by splitting multiple responsibilities into separate, well-defined files while preserving all detailed docstrings.
✅ Verification: Ran the ruff check and format, as well as the test suite for packages/workhorse and ensured 100% pass rate.
✨ Result: The codebase is cleaner and better modularized without any feature, documentation, or behavioral changes.
