+++ packages/workhorse/src/settler_workhorse/plugins/exceptions.py
@@ -1,0 +1,19 @@
+"""Plugin exceptions."""
+
+
+class PluginError(Exception):
+    """Base exception for plugin errors."""
+
+    pass
+
+
+class PluginLoadError(PluginError):
+    """Error loading a plugin."""
+
+    pass
+
+
+class HookError(PluginError):
+    """Error in hook execution."""
+
+    pass
